resource/mq_broker: Move expanders/flatteners to resource file
